Shift Nav to App.ks
