+++ frontend/App.js
@@ -1,20 +1,203 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import React, { useState } from "react";
+import {
+  StyleSheet,
+  Text,
+  View,
+  TouchableOpacity,
+  Modal,
+  Button,
+} from "react-native";
 import SignUpScreen from './screens/SignUpScreen';
 import LoginScreen from './screens/LoginScreen';
 import HomeScreen from './screens/HomeScreen';
+import Icon from 'react-native-vector-icons/Ionicons';
+import QRCode from 'react-native-qrcode-svg';
+import StoresScreen from './screens/StoresScreen';
+import RewardsScreen from './screens/RewardsScreen';
+import AccountScreen from './screens/AccountScreen';
+
+// const Stack = createNativeStackNavigator();
+
+// const App = () => {
+//   return (
+//     <NavigationContainer>
+//       <Stack.Navigator initialRouteName="SignUp">
+//         <Stack.Screen name="Sign Up" component={SignUpScreen} />
+//         <Stack.Screen name="Login" component={LoginScreen} />
+//         <Stack.Screen name="Home" component={HomeScreen} options={{ headerShown: false }} />
+//       </Stack.Navigator>
+//     </NavigationContainer>
+//   );
+// };
+// export default App;
 
 const Stack = createNativeStackNavigator();
+const Tab = createBottomTabNavigator();
 
-const App = () => {
+const CustomTabBar = ({ state, descriptors, navigation }) => {
+  const [modalVisible, setModalVisible] = React.useState(false);
+
+  return (
+    <View style={styles.tabBarContainer}>
+      {state.routes.map((route, index) => {
+        const { options } = descriptors[route.key];
+        const isFocused = state.index === index;
+
+        const onPress = () => {
+          if (route.name === 'QRButton') {
+            setModalVisible(true);
+          } else {
+            const event = navigation.emit({
+              type: 'tabPress',
+              target: route.key,
+              canPreventDefault: true,
+            });
+
+            if (!isFocused && !event.defaultPrevented) {
+              navigation.navigate(route.name);
+            }
+          }
+        };
+
+        const onLongPress = () => {
+          navigation.emit({
+            type: 'tabLongPress',
+            target: route.key,
+          });
+        };
+
+        const iconName = options.tabBarIcon ? options.tabBarIcon(isFocused) : 'circle';
+
+        return (
+          <TouchableOpacity
+            key={route.name}
+            accessibilityRole="button"
+            accessibilityState={isFocused ? { selected: true } : {}}
+            accessibilityLabel={options.tabBarAccessibilityLabel}
+            testID={options.tabBarTestID}
+            onPress={onPress}
+            onLongPress={onLongPress}
+            style={styles.tabBarItem}
+          >
+            {route.name === 'QRButton' ? (
+              <View style={styles.qrButton}>
+                <Icon name="qr-code" color="#fff" size={40} />
+              </View>
+            ) : (
+              <Icon name={iconName} color={isFocused ? '#88C34A' : '#222'} size={25} />
+            )}
+          </TouchableOpacity>
+        );
+      })}
+      <Modal
+        transparent={true}
+        visible={modalVisible}
+        animationType="slide"
+        onRequestClose={() => setModalVisible(false)}
+      >
+        <View style={styles.modalContainer}>
+          <View style={styles.modalContent}>
+            <QRCode value="https://example.com" size={200} />
+            <Button title="Close" onPress={() => setModalVisible(false)} />
+          </View>
+        </View>
+      </Modal>
+    </View>
+  );
+};
+
+const MainTabNavigator = () => (
+  <Tab.Navigator tabBar={props => <CustomTabBar {...props} />}>
+    <Tab.Screen
+      name="Home"
+      component={HomeScreen}
+      options={{ tabBarIcon: ({ focused }) => focused ? 'home' : 'home-outline' }}
+    />
+    <Tab.Screen
+      name="Stores"
+      component={StoresScreen}
+      options={{ tabBarIcon: ({ focused }) => focused ? 'storefront' : 'storefront-outline' }}
+    />
+    <Tab.Screen
+      name="QRButton"
+      component={View} // Placeholder component
+      options={{ tabBarButton: () => null }} // Hides the tab, button will be manually placed
+    />
+    <Tab.Screen
+      name="Rewards"
+      component={RewardsScreen}
+      options={{ tabBarIcon: ({ focused }) => focused ? 'gift' : 'gift-outline' }}
+    />
+    <Tab.Screen
+      name="Account"
+      component={AccountScreen}
+      options={{ tabBarIcon: ({ focused }) => focused ? 'person' : 'person-outline' }}
+    />
+  </Tab.Navigator>
+);
+
+export default function App() {
   return (
     <NavigationContainer>
       <Stack.Navigator initialRouteName="SignUp">
-        <Stack.Screen name="Sign Up" component={SignUpScreen} />
-        <Stack.Screen name="Login" component={LoginScreen} />
-        <Stack.Screen name="Home" component={HomeScreen} options={{ headerShown: false }} />
+        <Stack.Screen
+          name="SignUp"
+          component={SignUpScreen}
+          options={{ headerShown: false }}
+        />
+        <Stack.Screen
+          name="Login"
+          component={LoginScreen}
+          options={{ headerShown: false }}
+        />
+        <Stack.Screen
+          name="Main"
+          component={MainTabNavigator}
+          options={{ headerShown: false }}
+        />
       </Stack.Navigator>
     </NavigationContainer>
   );
-};
-export default App;
+}
+
+const styles = StyleSheet.create({
+  tabBarContainer: {
+    flexDirection: 'row',
+    height: 75,
+    borderTopWidth: 1,
+    borderTopColor: '#ccc',
+    backgroundColor: '#fff',
+    justifyContent: 'space-around',
+    alignItems: 'center',
+  },
+  tabBarItem: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  qrButton: {
+    position: 'absolute',
+    bottom: 10,
+    alignSelf: 'center',
+    backgroundColor: '#88C34A',
+    padding: 25,
+    borderRadius: 50,
+    zIndex: 10,
+  },
+  modalContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.5)',
+  },
+  modalContent: {
+    width: 300,
+    padding: 20,
+    backgroundColor: 'white',
+    borderRadius: 10,
+    alignItems: 'center',
+  },
+});
