+++ frontend/package.json
@@ -20,7 +20,8 @@
     "react-native-qrcode-svg": "^6.3.1",
     "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
-    "react-native-svg": "^15.3.0"
+    "react-native-svg": "^15.3.0",
+    "react-native-vector-icons": "^10.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
